requirements.txt
phidata
yfinance
packaging